Change "down" key to move piece to bottom of board
The full implementation of the "down" key was implemented. An extra
BoardUpdateResult (REACHED_BOTTOM) was added to represent when a piece
has reached the bottom of the board. When the "down" key is pressed
This is used to repetitively move a piece down on the board until a
state of REACHED_BOTTOM is reached.